+++ ex-mortis/appactivatable.py
@@ -224,7 +224,7 @@ class ExMortisAppActivatable(
 
 		window_manager.track_window(window)
 
-		self.setup_restore_window(window_manager, window)
+		self.setup_restore_window(window_manager, settings, window)
 
 		if self.is_saving_window_states():
 			self.bind_window_settings(window_manager, settings, window)